requirements/dev.txt
-r base.txt
django-debug-toolbar==2.2
black==20.8b1
flake8==3.8.3
pytest==6.0.1
pytest-django==3.9.0